Refactored project structure: - Got rid of the duplicate Course classes - Updated the build.app to fix an error that would not allow the program to build - Refactored other classes to use the new Course class that is combination of the two
