+++ app/build.gradle
@@ -33,6 +33,10 @@ android {
     }
 }
 
+configurations.all {
+    exclude group: "com.intellij", module: "annotations"
+}
+
 dependencies {
     implementation libs.appcompat
     implementation libs.material
